refactor: привёл в порядок пакеты
