+++ src/main/java/ru/practicum/shareit/booking/model/Booking.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.booking;
+package ru.practicum.shareit.booking.model;
 
 /**
  * TODO Sprint add-bookings.
