+++ src/main/java/ru/practicum/shareit/user/dto/UserDto.java
@@ -1,0 +1,4 @@
+package ru.practicum.shareit.user.dto;
+
+public class UserDto {
+}
